# Optimize newline handling for RegexOptions.Multiline