@danmosemsft 2020-04-06
Another case to handle if @jzabroski still plans to revive https://github.com/dotnet/corefx/pull/41195