@jzabroski 2020-04-07
@danmosemsft I do plan to.   Just got trapped under mountain of COVID-19 DR planning work.  I expect to give it the old college try in about 2 weeks.  Thanks for thinking of me and tagging me.